CR remarks
