+++ visual-espresso/gradle/libs.versions.toml
@@ -8,6 +8,7 @@ espressoCore = "3.6.1"
 appcompat = "1.7.0"
 material = "1.12.0"
 monitor = "1.7.2"
+visualJava = "0.11.0"
 
 [libraries]
 httpclient = { module = "cz.msebera.android:httpclient", version.ref = "httpclient" }
@@ -18,6 +19,7 @@ espresso-core = { group = "androidx.test.espresso", name = "espresso-core", vers
 appcompat = { group = "androidx.appcompat", name = "appcompat", version.ref = "appcompat" }
 material = { group = "com.google.android.material", name = "material", version.ref = "material" }
 monitor = { group = "androidx.test", name = "monitor", version.ref = "monitor" }
+visualJava = { group = "com.saucelabs.visual", name = "java-client", version.ref = "visualJava"}
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
