+++ visual-espresso/visual/build.gradle
@@ -43,6 +43,7 @@ dependencies {
     implementation libs.material
     implementation libs.monitor
     implementation libs.junit
+    implementation libs.visualJava
     androidTestImplementation libs.ext.junit
     androidTestImplementation libs.espresso.core
 }
