--- visual-espresso/visual/src/main/java/com/saucelabs/visual/graphql/GraphQLOperation.java
@@ -1,9 +1,0 @@
-package com.saucelabs.visual.graphql;
-
-import java.util.Map;
-
-public interface GraphQLOperation {
-    String getQuery();
-
-    Map<String, Object> getVariables();
-}
