+++ visual-espresso/visual/src/main/java/com/saucelabs/visual/junit/TestMetaInfoRule.java
